[#32]refactor: 무한 스크롤 커스컴 훅스 작성 및 적용
IntersectionObserver 관련 코드를 커스텀 훅스로 작성
